src/main/java/br/ufpb/Biblioteca_API/Services/AuthorService.java
package br.ufpb.Biblioteca_API.Services;

import br.ufpb.Biblioteca_API.Exception.AuthorNotFoundException;
import br.ufpb.Biblioteca_API.Models.Author;
import br.ufpb.Biblioteca_API.Repositories.AuthorRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.List;

@Service
public class AuthorService {

    @Autowired
    private AuthorRepository authorRepository;

    public List<Author> getAllAuthors() {
        return authorRepository.findAll();
    }

    public Author getAuthorById(Long id) {
        return authorRepository.findById(id)
                .orElseThrow(() -> new AuthorNotFoundException("Author with ID " + id + " not found."));
    }

    public Author saveAuthor(Author author) {
        return authorRepository.save(author);
    }

    public Author updateAuthor(Long id, Author author) {
        Author existingAuthor = authorRepository.findById(id)
                .orElseThrow(() -> new AuthorNotFoundException("Author with ID " + id + " not found."));
        author.setId(existingAuthor.getId());
        return authorRepository.save(author);
    }

    public void deleteAuthor(Long id) {
        Author existingAuthor = authorRepository.findById(id)
                .orElseThrow(() -> new AuthorNotFoundException("Author with ID " + id + " not found."));
        authorRepository.delete(existingAuthor);
    }
}
